asdasdasd

+++ src/main/java/org/example/repository/AlumnoRepository.java
@@ -3,5 +3,6 @@ package org.example.repository;
 import org.example.model.Alumno;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AlumnoRepository extends JpaRepository<Alumno, Long> {
+public interface AlumnoRepository extends JpaRepository<Alumno, Integer> {
+
 }
